mapping implemented

+++ main/controllers/user_controller.py
@@ -10,6 +10,7 @@ class UserController(Resource):
     user_req_valid_args.add_argument("id", type=int)
     user_req_valid_args.add_argument("name", type=str, required=True)
     user_req_valid_args.add_argument("email", type=str, required=True)
+    user_req_valid_args.add_argument("date_created", type=str)
 
     def get(self):
         query_parameters = request.args
@@ -22,15 +23,7 @@ class UserController(Resource):
 
     def post(self):
         new_user = self.user_req_valid_args.parse_args()
-
-        result = self.user_repository.save(new_user)
-
-        if result is None:
-            print("user save failed")
-            return "user save failed"
-        else:
-            print("user saved successfully")
-            return result
+        return self.user_repository.save(new_user)
 
     def put(self):
         update_user_req_valid = self.user_req_valid_args.copy()
@@ -39,16 +32,13 @@ class UserController(Resource):
         new_data = update_user_req_valid.parse_args()
 
         result = self.user_repository.update(new_data["id"], new_data)
-
-        if result is None:
-            print("user update failed")
-            return "user update failed"
-        else:
-            print("user updated successfully")
-            return result
+        return result
 
     def delete(self):
         query_parameters = request.args
         user_id = query_parameters.get('id')
 
-        return self.user_repository.delete(int(user_id))
+        if user_id:
+            return self.user_repository.delete(int(user_id))
+        else:
+            return False
